requirements.txt
streamlit==1.29.0
pdf2image
pytesseract
numpy
streamlit-drawable-canvas==0.9.3
pandas
Pillow